requirements.txt
psycopg2-binary==2.9.6
python-dotenv==1.0.0
faker==33.1.0
